+++ src/components/structuredlang/Tuple.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+const Tuple = ({ elements }) => (
+    <span>Tuple: NYI</span>
+)
+
+Tuple.propTypes = {
+    elements: PropTypes.array.isRequired,
+};
+
+export default Tuple;
